refactor script structure
