+++ .github/workflows/prepare-prerelease.yml
@@ -70,7 +70,7 @@ jobs:
         id: check-frozen
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: ./node_modules/.bin/ts-node --swc ./release/is-pr-frozen.ts
+        run: yarn release:is-pr-frozen
 
       - name: Cancel when frozen
         if: steps.check-frozen.outputs.frozen == 'true' && github.event_name != 'workflow_dispatch'
