+++ scripts/package.json
@@ -12,9 +12,16 @@
     "lint:js:cmd": "cross-env NODE_ENV=production eslint --cache --cache-location=../.cache/eslint --ext .js,.jsx,.json,.html,.ts,.tsx,.mjs --report-unused-disable-directives",
     "lint:package": "sort-package-json",
     "migrate-docs": "node --require esbuild-register ./ts-to-ts49.ts",
+    "release:generate-pr-description": "ts-node --swc ./release/generate-pr-description.ts",
+    "release:get-changelog-from-file": "ts-node --swc ./release/get-changelog-from-file.ts",
+    "release:get-current-version": "ts-node --swc ./release/get-current-version.ts",
+    "release:get-version-changelog": "ts-node --swc ./release/get-version-changelog.ts",
+    "release:is-pr-frozen": "ts-node --swc ./release/is-pr-frozen.ts",
+    "release:is-prerelease": "ts-node --swc ./release/is-prerelease.ts",
+    "release:is-version-published": "ts-node --swc ./release/is-version-published.ts",
+    "release:publish": "ts-node --swc ./release/publish.ts",
     "release:version": "ts-node --swc ./release/version.ts",
     "release:write-changelog": "ts-node --swc ./release/write-changelog.ts",
-    "release:publish": "ts-node --swc ./release/publish.ts",
     "task": "ts-node --swc ./task.ts",
     "test": "jest --config ./jest.config.js",
     "upgrade": "ts-node --swc ./task.ts"
