+++ scripts/release/get-current-version.ts
